server/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.1.2
cffi==2.0.0
click==8.3.1
cryptography==43.0.0
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.0
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==2.4.3
pandas==2.2.2
pillow==10.4.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.9
pytz==2026.1.post1
PyYAML==6.0.3
qrcode==7.4.2
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.31
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0
